CNN code for Lesson 8
I mistakenly named the L7 code lesson8.py in the previous commit. I updated the naming to lesson7.py, and also added lesson8.py code which is adds CNN development code to the end of the lesson7.py code, so you could compare the performance of MLP and CNN for MNIST classification problem.